requirements.txt
streamlit>=1.32.0
google-generativeai>=0.8.0
libsql-experimental>=0.0.11
